Move `sources_finished` into `impl TopologyController`
